requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
google-genai
mistralai
